+++ lib/src/builder.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:build/build.dart';
 import 'package:fast_i18n/src/generator/generate.dart';
+import 'package:fast_i18n/src/model/build_config.dart';
 import 'package:fast_i18n/src/model/i18n_config.dart';
 import 'package:fast_i18n/src/model/i18n_data.dart';
 import 'package:fast_i18n/src/model/i18n_locale.dart';
@@ -15,38 +16,39 @@ Builder i18nBuilder(BuilderOptions options) => I18nBuilder(options);
 class I18nBuilder implements Builder {
   I18nBuilder(this.options);
 
-  static const String defaultBaseLocale = 'en';
-  static const String defaultBaseName = 'strings';
-  static const String defaultTranslateVar = 't';
-  static const String defaultEnumName = 'AppLocale';
-  static const String defaultInputFilePattern = '.i18n.json';
-  static const String defaultOutputFilePattern = '.g.dart';
-
   final BuilderOptions options;
 
   bool _generated = false;
 
   String get inputFilePattern =>
-      options.config['input_file_pattern'] ?? defaultInputFilePattern;
+      options.config['input_file_pattern'] ??
+      BuildConfig.defaultInputFilePattern;
   String get outputFilePattern =>
-      options.config['output_file_pattern'] ?? defaultOutputFilePattern;
+      options.config['output_file_pattern'] ??
+      BuildConfig.defaultOutputFilePattern;
 
   @override
   FutureOr<void> build(BuildStep buildStep) async {
-    final I18nLocale baseLocale = I18nLocale.fromString(
-        options.config['base_locale'] ?? defaultBaseLocale);
-    final String? inputDirectory = options.config['input_directory'];
-    final String? outputDirectory = options.config['output_directory'];
-    final String translateVar =
-        options.config['translate_var'] ?? defaultTranslateVar;
-    final String enumName = options.config['enum_name'] ?? defaultEnumName;
-    final String? translationClassVisibility =
-        options.config['translation_class_visibility'];
-    final String? keyCase = options.config['key_case'];
-    final List<String> maps = options.config['maps']?.cast<String>() ?? [];
+    final buildConfig = BuildConfig(
+        baseLocale: I18nLocale.fromString(
+            options.config['base_locale'] ?? BuildConfig.defaultBaseLocale),
+        inputDirectory: options.config['input_directory'],
+        inputFilePattern: inputFilePattern,
+        outputDirectory: options.config['output_directory'],
+        outputFilePattern: outputFilePattern,
+        translateVar:
+            options.config['translate_var'] ?? BuildConfig.defaultTranslateVar,
+        enumName: options.config['enum_name'] ?? BuildConfig.defaultEnumName,
+        translationClassVisibility:
+            (options.config['translation_class_visibility'] as String?)
+                    .toTranslationClassVisibility() ??
+                BuildConfig.defaultTranslationClassVisibility,
+        keyCase: (options.config['key_case'] as String?).toKeyCase() ??
+            BuildConfig.defaultKeyCase,
+        maps: options.config['maps']?.cast<String>() ?? []);
 
-    if (inputDirectory != null &&
-        !buildStep.inputId.path.contains(inputDirectory)) return;
+    if (buildConfig.inputDirectory != null &&
+        !buildStep.inputId.path.contains(buildConfig.inputDirectory!)) return;
 
     // only generate once
     if (_generated) return;
@@ -57,8 +59,8 @@ class I18nBuilder implements Builder {
     final Map<AssetId, I18nLocale> assetMap = Map();
     String? baseName;
 
-    final Glob findAssetsPattern = inputDirectory != null
-        ? Glob('**$inputDirectory/*$inputFilePattern')
+    final Glob findAssetsPattern = buildConfig.inputDirectory != null
+        ? Glob('**${buildConfig.inputDirectory}/*$inputFilePattern')
         : Glob('**$inputFilePattern');
 
     await buildStep.findAssets(findAssetsPattern).forEach((assetId) {
@@ -68,7 +70,7 @@ class I18nBuilder implements Builder {
       final baseFile = Utils.baseFileRegex.firstMatch(fileNameNoExtension);
       if (baseFile != null) {
         // base file
-        assetMap[assetId] = baseLocale;
+        assetMap[assetId] = buildConfig.baseLocale;
         baseName = fileNameNoExtension;
       } else {
         // secondary files (strings_x)
@@ -83,19 +85,10 @@ class I18nBuilder implements Builder {
       }
     });
 
-    if (baseName == null) baseName = defaultBaseName;
-
-    // build config which applies to all locales
-    final config = I18nConfig(
-        baseName: baseName!,
-        baseLocale: baseLocale,
-        maps: maps,
-        keyCase: keyCase.toKeyCase(),
-        translateVariable: translateVar,
-        enumName: enumName,
-        translationClassVisibility:
-            translationClassVisibility.toTranslationClassVisibility() ??
-                TranslationClassVisibility.private);
+    if (baseName == null) {
+      print('Error: No base translation file.');
+      return;
+    }
 
     // map each assetId to I18nData
     final localesWithData = Map<AssetId, I18nData>();
@@ -103,13 +96,19 @@ class I18nBuilder implements Builder {
     for (MapEntry<AssetId, I18nLocale> asset in assetMap.entries) {
       I18nLocale locale = asset.value;
       String content = await buildStep.readAsString(asset.key);
-      I18nData representation = parseJSON(config, locale, content);
+      I18nData representation = parseJSON(buildConfig, locale, content);
       localesWithData[asset.key] = representation;
     }
 
     // generate
     final String output = generate(
-        config: config,
+        config: I18nConfig(
+            baseName: baseName!,
+            baseLocale: buildConfig.baseLocale,
+            keyCase: buildConfig.keyCase,
+            translateVariable: buildConfig.translateVar,
+            enumName: buildConfig.enumName,
+            translationClassVisibility: buildConfig.translationClassVisibility),
         translations: localesWithData.values.toList()
           ..sort((a, b) => a.base
               ? -1
@@ -120,8 +119,8 @@ class I18nBuilder implements Builder {
     final AssetId baseId =
         localesWithData.entries.firstWhere((element) => element.value.base).key;
 
-    final finalOutputDirectory =
-        outputDirectory ?? (baseId.pathSegments..removeLast()).join('/');
+    final finalOutputDirectory = buildConfig.outputDirectory ??
+        (baseId.pathSegments..removeLast()).join('/');
     final String outFilePath =
         '$finalOutputDirectory/$baseName$outputFilePattern';